wip: refactor modules
