+++ DesignSystemModule/Sources/Extensions/Enum/TransactionType+Extensions.swift
@@ -1,0 +1,22 @@
+//
+//  File.swift
+//  DesignSystemModule
+//
+//  Created by Theo Sementa on 27/08/2025.
+//
+
+import Foundation
+import Models
+import CoreModule
+
+public extension TransactionType {
+  
+  var name: String {
+      switch self {
+      case .expense:  return Word.Classic.expense
+      case .income:   return Word.Classic.income
+      case .transfer: return Word.Main.transfer
+      }
+  }
+  
+}
