+++ Mocks/Sources/SubcategoryMock.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import Models
 
 public extension SubcategoryModel {
     static let mock = SubcategoryModel(
