+++ Models/Package.swift
@@ -4,22 +4,22 @@
 import PackageDescription
 
 let package = Package(
-    name: "Navigation",
+    name: "Models",
     platforms: [.iOS(.v17)],
     products: [
         // Products define the executables and libraries a package produces, making them visible to other packages.
         .library(
-            name: "Navigation",
-            targets: ["Navigation"]),
+            name: "Models",
+            targets: ["Models"]),
     ],
     targets: [
         // Targets are the basic building blocks of a package, defining a module or a test suite.
         // Targets can depend on other targets in this package and products from dependencies.
         .target(
-            name: "Navigation"),
+            name: "Models"),
         .testTarget(
-            name: "NavigationTests",
-            dependencies: ["Navigation"]
+            name: "ModelsTests",
+            dependencies: ["Models"]
         ),
     ]
 )
