+++ Models/Sources/Enum/AccountType.swift
@@ -1,0 +1,13 @@
+//
+//  AccountType.swift
+//  Models
+//
+//  Created by Theo Sementa on 27/08/2025.
+//
+
+import Foundation
+
+public enum AccountType: Int, CaseIterable {
+    case classic = 0
+    case savings = 1
+}
