+++ Models/Sources/Enum/TransactionType.swift
@@ -1,0 +1,14 @@
+//
+//  TransactionType.swift
+//  CashFlow
+//
+//  Created by Theo Sementa on 30/04/2025.
+//
+
+import Foundation
+
+public enum TransactionType: Int, CaseIterable {
+    case expense = 0
+    case income = 1
+    case transfer = 2
+}
